pyproject.toml
[project]
name = "medgraph-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.86",
    "docling>=2.68.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=4.2.0",
    "neo4j>=5.20.0",
    "pubmed2pdf>=0.0.7",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "ragas>=0.2.0",
    "langgraph>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-google-vertexai>=2.0.0",
]
